Remove ReportPoValue entity
This data is now captured by the Report entity.
Remove the report_po_value table.
