+++ src/Ilios/CoreBundle/Entity/Report.php
@@ -84,6 +84,20 @@ class Report implements ReportInterface
     protected $prepositionalObject;
 
     /**
+     * @var string
+     *
+     * @ORM\Column(name="prepositional_object_table_row_id", type="string", length=14, nullable=true)
+     *
+     * @Assert\NotBlank()
+     * @Assert\Type(type="string")
+     * @Assert\Length(
+     *      min = 1,
+     *      max = 14
+     * )
+     */
+    protected $prepositionalObjectTableRowId;
+
+    /**
      * @var boolean
      *
      *@ORM\Column(name="deleted", type="boolean")
@@ -104,17 +118,6 @@ class Report implements ReportInterface
      * @JMS\Type("string")
      */
     protected $user;
-
-    /**
-    * @var ReportPoValueInterface
-    *
-    * @ORM\OneToOne(targetEntity="ReportPoValue", mappedBy="report")
-    *
-    * @JMS\Expose
-    * @JMS\Type("string")
-    * @JMS\SerializedName("poValue")
-    */
-    protected $poValue;
 
     /**
      * Default createdAt to now
@@ -187,6 +190,22 @@ class Report implements ReportInterface
     public function getPrepositionalObject()
     {
         return $this->prepositionalObject;
+    }
+
+    /**
+     * @param string $prepositionalObjectTableRowId
+     */
+    public function setPrepositionalObjectTableRowId($prepositionalObjectTableRowId)
+    {
+        $this->prepositionalObjectTableRowId = $prepositionalObjectTableRowId;
+    }
+
+    /**
+     * @return string
+     */
+    public function getPrepositionalObjectTableRowId()
+    {
+        return $this->prepositionalObjectTableRowId;
     }
 
     /**
