+++ src/Ilios/CoreBundle/Resources/config/handlers.yml
@@ -42,7 +42,6 @@ parameters:
     ilioscore.publishevent.handler.class: Ilios\CoreBundle\Handler\PublishEventHandler
     ilioscore.recurringevent.handler.class: Ilios\CoreBundle\Handler\RecurringEventHandler
     ilioscore.report.handler.class: Ilios\CoreBundle\Handler\ReportHandler
-    ilioscore.reportpovalue.handler.class: Ilios\CoreBundle\Handler\ReportPoValueHandler
     ilioscore.school.handler.class: Ilios\CoreBundle\Handler\SchoolHandler
     ilioscore.sessiondescription.handler.class: Ilios\CoreBundle\Handler\SessionDescriptionHandler
     ilioscore.sessionlearningmaterial.handler.class: Ilios\CoreBundle\Handler\SessionLearningMaterialHandler
@@ -185,9 +184,6 @@ services:
     ilioscore.report.handler:
         class: %ilioscore.report.handler.class%
         arguments: ['@doctrine', Ilios\CoreBundle\Entity\Report, '@form.factory']
-    ilioscore.reportpovalue.handler:
-        class: %ilioscore.reportpovalue.handler.class%
-        arguments: ['@doctrine', Ilios\CoreBundle\Entity\ReportPoValue, '@form.factory']
     ilioscore.school.handler:
         class: %ilioscore.school.handler.class%
         arguments: ['@doctrine', Ilios\CoreBundle\Entity\School, '@form.factory']
