+++ src/Ilios/CoreBundle/Resources/config/routing.yml
@@ -246,14 +246,6 @@ api_report_v1:
     prefix:   /v1
     defaults: {_format:json}
 
-
-api_reportpovalue_v1:
-    resource: "@IliosCoreBundle/Controller/ReportPoValueController.php"
-    type:     rest
-    prefix:   /v1
-    defaults: {_format:json}
-
-
 api_school_v1:
     resource: "@IliosCoreBundle/Controller/SchoolController.php"
     type:     rest
