--- src/Ilios/CoreBundle/Tests/DataLoader/ReportPoValueData.php
@@ -1,24 +1,0 @@
-<?php
-
-namespace Ilios\CoreBundle\Tests\DataLoader;
-
-class ReportPoValueData extends AbstractDataLoader
-{
-    protected function getData()
-    {
-        $arr = array();
-
-
-        return $arr;
-    }
-
-    public function create()
-    {
-        return [];
-    }
-
-    public function createInvalid()
-    {
-        return [];
-    }
-}
