--- src/Ilios/CoreBundle/Tests/Fixture/LoadReportPoValueData.php
@@ -1,38 +1,0 @@
-<?php
-
-namespace Ilios\CoreBundle\Tests\Fixture;
-
-use Ilios\CoreBundle\Entity\ReportPoValue;
-use Doctrine\Common\DataFixtures\AbstractFixture;
-use Doctrine\Common\DataFixtures\FixtureInterface;
-use Doctrine\Common\Persistence\ObjectManager;
-use Symfony\Component\DependencyInjection\ContainerAwareInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
-
-class LoadReportPoValueData extends AbstractFixture implements
-    FixtureInterface,
-    ContainerAwareInterface
-{
-
-    private $container;
-
-    public function setContainer(ContainerInterface $container = null)
-    {
-        $this->container = $container;
-    }
-
-    public function load(ObjectManager $manager)
-    {
-        $data = $this->container
-            ->get('ilioscore.dataloader.reportPoValue')
-            ->getAll();
-        foreach ($data as $arr) {
-            $entity = new ReportPoValue();
-            $entity->setId($arr['id']);
-            $manager->persist($entity);
-            $this->addReference('reportPoValues' . $arr['id'], $entity);
-        }
-
-        $manager->flush();
-    }
-}
